Allow custom host name value
