+++ stagemonitor-core/src/main/java/org/stagemonitor/core/MeasurementSession.java
@@ -118,24 +118,4 @@ public class MeasurementSession {
 		return stringRepresentation;
 	}
 
-	public static String getNameOfLocalHost() {
-		try {
-			return InetAddress.getLocalHost().getHostName();
-		} catch (Exception e) {
-			return getHostNameFromEnv();
-		}
-	}
-
-	static String getHostNameFromEnv() {
-		// try environment properties.
-		String host = System.getenv("COMPUTERNAME");
-		if (host != null) {
-			return host;
-		}
-		host = System.getenv("HOSTNAME");
-		if (host != null) {
-			return host;
-		}
-		return null;
-	}
 }
